componentes reutilizables nav, inputList, perfilLiquidador.
