+++ pages/jefeLiquidador/liquidadoresViews/navData.tsx
@@ -1,0 +1,27 @@
+
+import WarningIcon from "../../../libs/components/icons/AlertHandDrawnSymbol.component";
+import PresuIcon from "../../../libs/components/icons/FoldedNewspaper.component";
+import UserIcon from "../../../libs/components/icons/BaselineFace24px.component";
+
+
+
+export default [
+    {
+        url: "/jefeLiquidador/liquidadoresViews/perfilLiquidador",
+        icon: <UserIcon />,
+        txt: "liquidador",
+
+    },
+    {
+        url: "/jefeLiquidador/liquidadoresViews/siniestro",
+        icon: <WarningIcon />,
+        txt: "siniestro",
+
+    },
+    {
+        url: "/jefeLiquidador/liquidadoresViews/presupuestos",
+        icon: <PresuIcon />,
+        txt: "presupuesto",
+
+    }
+  ]
