+++ pages/jefeLiquidador/liquidadoresViews/perfilLiquidador/index.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import styled from "styled-components";
 import LayoutViews from "../../../../libs/components/layoutViews";
 import NavViews from "../../../../libs/components/navViews";
 import navData from "../navData";
@@ -9,22 +8,14 @@ export default () => {
     return (
 
         <LayoutViews user="jefe">
-            <DirectionCenter>
-                <NavViews data={navData} />
+           <NavViews data={navData} classCss={"max-width:350px;margin:10px auto 15px auto;"} />
+    
                 <PerfilLiquidador info={info} />
-            </DirectionCenter>
+         
         </LayoutViews>
-
     )
 }
 
-const DirectionCenter = styled.div`
-    width:90%;
-    margin:12px auto;
-    @media(max-width:780px){
-      width:100%;
-    }
-`;
 
 const info = {
     name: "juan jose araneda",
