+++ pages/jefeLiquidador/siniestrosViews/navData.tsx
@@ -1,0 +1,63 @@
+
+import SiniestroIcon from "../../../libs/components/icons/AlertHandDrawnSymbol.component";
+import TrabajoIcon from "../../../libs/components/icons/CustomerSupport.component";
+import UserIcon from "../../../libs/components/icons/BaselineFace24px.component";
+import PolizaIcon from "../../../libs/components/icons/Insurance.component";
+import VehiculoIcon from "../../../libs/components/icons/CarOfHatchback.component";
+import TallerIcon from "../../../libs/components/icons/Carmechanic89096.component";
+import PresupuestoIcon from "../../../libs/components/icons/FoldedNewspaper.component";
+import MultimediaIcon from "../../../libs/components/icons/Gallery.component";
+
+
+
+
+export default [
+    {
+        url: "/jefeLiquidador/siniestrosViews/perfilLiquidador",
+        icon: <UserIcon />,
+        txt: "liquidador",
+  
+    },
+    {
+        url: "",
+        icon: <PolizaIcon />,
+        txt: "poliza",
+  
+    },
+    {
+        url: "/jefeLiquidador/siniestrosViews/siniestro",
+        icon: <SiniestroIcon />,
+        txt: "siniestro",
+  
+    },
+    {
+        url: "",
+        icon: <VehiculoIcon />,
+        txt: "vehículos",
+  
+    },
+    {
+        url: "",
+        icon: <TrabajoIcon />,
+        txt: "trabajos",
+  
+    },
+    {
+        url: "/jefeLiquidador/siniestrosViews/taller",
+        icon: <TallerIcon />,
+        txt: "taller",
+  
+    },
+    {
+        url: "",
+        icon: <PresupuestoIcon />,
+        txt: "presupuesto",
+  
+    },
+    {
+        url: "",
+        icon: <MultimediaIcon />,
+        txt: "multimedia",
+  
+    }
+  ]
